+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add method `disableExpiration` to the class `PublishPress\Future\Modules\Expirator\Models\ExpirablePostModel`.
 
+### Changed
+
+- Move notification settings to a specific tab (Issue #190).
+
 ### Fixed
 
 - Do not remove expiration post meta when clearing the scheduled action (Issue #1053).